refactor: criado pasta utils para armazenar funcoes
