+++ src/views/Calculator.jsx
@@ -1,30 +1,11 @@
 import React, { useEffect, useState } from 'react';
+import { calculateWallArea } from '../utils/calculatorFuncions';
 
 const wallParameters = {
   width: '0',
   height: '0',
   doors: '0',
   windows: '0',
-};
-
-const predefinedAreas = {
-  doorArea: (0.80 * 1.9),
-  windowArea: (2 * 1.2),
-};
-
-// const wallPaintOptions = [0.5, 2.5, 3.6, 18];
-
-const calculateWallArea = ({
-  width, height, doors, windows,
-}) => {
-  const windowsArea = windows > 0 ? predefinedAreas.windowArea * windows : 1;
-  const doorsArea = doors > 0 ? predefinedAreas.doorArea * doors : 1;
-
-  const wallArea = width * height;
-
-  const areaToBePainted = wallArea - windowsArea - doorsArea;
-
-  return areaToBePainted;
 };
 
 function Calculator() {
